chore: folder structure changes
