+++ lib/sqlite3_common.dart
@@ -1,0 +1,2 @@
+// Exports common Sqlite3 exports which are available in different environments.
+export 'package:sqlite3/common.dart';
