+++ lib/src/impl/isolate_connection_factory_impl.dart
@@ -1,0 +1,3 @@
+export 'stub_isolate_connection_factory.dart'
+    // ignore: uri_does_not_exist
+    if (dart.library.io) '../native/native_isolate_connection_factory.dart';
