+++ lib/src/impl/open_factory_impl.dart
@@ -1,0 +1,3 @@
+export 'stub_sqlite_open_factory.dart'
+    // ignore: uri_does_not_exist
+    if (dart.library.io) '../native/native_sqlite_open_factory.dart';
