+++ lib/src/utils/native_database_utils.dart
@@ -1,0 +1,13 @@
+import 'dart:isolate';
+
+class SubscribeToUpdates {
+  final SendPort port;
+
+  SubscribeToUpdates(this.port);
+}
+
+class UnsubscribeToUpdates {
+  final SendPort port;
+
+  UnsubscribeToUpdates(this.port);
+}
