+++ test/utils/test_utils_impl.dart
@@ -1,0 +1,3 @@
+export 'stub_test_utils.dart'
+    // ignore: uri_does_not_exist
+    if (dart.library.io) 'native_test_utils.dart';
